+++ classes/app.rb
