+++ classes/base_decorator.rb
